// src/utils/types/user.ts
export type User = {
  _id: string;
  username: string;
  email: string;
  password?: string;
  firstName: string;
  lastName: string;
  pfp?: string;
};

export type GivenUser = {
  _id: string;
  username: string;
  firstName: string;
  lastName: string;
  email?: string;
  pfp?: string;
};

export type PossibleUser = PublicUser | AutoGeneratedUser;

export interface PublicUser extends UserFragment {
  username: string;
  email?: string;
  pfp?: string;
  bio?: string;
  type: 1;
}

export interface AutoGeneratedUser extends UserFragment {

  type: 0;
}
export interface UserFragment {
  _id: string;
  firstName: string;
  lastName: string;
  type: 0 | 1; // 0 = unverified, 1 = verified
}

export type UnverifiedUser = {
  _id: string;
  username: string;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
};